Repo cleanup - Rewards (#77)
For now did not touch maps and filters, they should be extracted and named properly for clearity.

Cleaned up Rewards, all logic is untouched, only organized in RewardsContext.
Moved it as seperate feature, components should be simple reusable fuctions.
Please use named imports from now on, and we'll refactor as we go.

This is generally how logic should be separated from the txs structure, with context and hooks.

- Refactored Rewards logic into RewardsContext, fixed prop drilling [official docs](https://react.dev/learn/passing-data-deeply-with-context)
- moved ComposeReward from Landing to features/Rewards [react project structure](https://dev.to/pramod_boda/recommended-folder-structure-for-react-2025-48mc)
- renamed Reward Claim to Claimed Rewards
- renamed AddEvents to Events
- used named imports instead of default